qag_paper/requirements.txt
torch==2.3.0+cu121
torchvision==0.18.0+cu121
transformers==4.40.2
bitsandbytes==0.43.1
accelerate==0.30.0
sentencepiece==0.2.0
protobuf==4.25.3
datasets==2.19.1
evaluate==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
jupyter==1.0.0
ipykernel==6.29.4
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.0.1
openai==1.30.1
tiktoken==0.7.0
llama-cpp-python==0.2.77
huggingface_hub==0.23.1

# Optional for quantization-scheme comparison experiments
optimum==1.21.4
# NOTE: GPTQ/AWQ packages are environment-specific and may conflict with this pinned stack.
# Install them in a separate env when running scheme-comparison experiments.
